액티비티 전환 수정
