+++ app/src/main/java/campus/tech/kakao/map/viewmodel/MyViewModel.kt
@@ -33,16 +33,13 @@ class MyViewModel(private val context: Context, private val repository: MyReposi
 
     val vmPlaceAdapter: PlaceAdapter = PlaceAdapter(listOf()) { place ->  //리사이클러뷰의 아이템을 누르면
         repository.insertSavedsearch(place.id, place.name)
-        savedSearchAdapterUpdateData.value = repository.getSavedSearches()
         updateSavedSearch()
-        setSharedPreferences(place)
         itemClick.value = place
     }
 
     val vmSavedSearchAdapter: SavedSearchAdapter = SavedSearchAdapter(listOf(),
         onCloseClick = { SavedSearch -> //SavedSearch의 x를 누르면
             repository.deleteSavedSearch(SavedSearch.id)
-            savedSearchAdapterUpdateData.value = repository.getSavedSearches()
             updateSavedSearch()
         },
         onNameClick = { SavedSearch ->   //SavedSearch의 이름을 누르면
@@ -51,19 +48,19 @@ class MyViewModel(private val context: Context, private val repository: MyReposi
         }
     )
 
-    fun setSharedPreferences(place: Place) {
-        //sharedPreference를 이용해서 name,address,latitude,longitude 저장하기
-        val sharedPreferences =context.getSharedPreferences(
-            "PlacePreferences",
-            AppCompatActivity.MODE_PRIVATE
-        )
-        val editor = sharedPreferences.edit()
-        editor.putString("name", place.name)
-        editor.putString("address", place.address)
-        editor.putString("longitude", place.longitude)
-        editor.putString("latitude", place.latitude)
-        editor.apply()
-    }
+//    fun setSharedPreferences(place: Place) {
+//        //sharedPreference를 이용해서 name,address,latitude,longitude 저장하기
+//        val sharedPreferences =context.getSharedPreferences(
+//            "PlacePreferences",
+//            AppCompatActivity.MODE_PRIVATE
+//        )
+//        val editor = sharedPreferences.edit()
+//        editor.putString("name", place.name)
+//        editor.putString("address", place.address)
+//        editor.putString("longitude", place.longitude)
+//        editor.putString("latitude", place.latitude)
+//        editor.apply()
+//    }
 
 
     fun intentSearchPlace() {   // true일 때 SearchPlaceActivity
